requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via discord-py
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
charset-normalizer==2.0.12
    # via aiohttp
discord.py @ git+https://github.com/rapptz/discord.py.git
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
pydantic[dotenv]==1.9.0
    # via -r requirements.in
python-dotenv==0.19.2
    # via pydantic
typing-extensions==4.1.1
    # via pydantic
yarl==1.7.2
    # via aiohttp
